# Faster page load: delay rendering codemirror until in view