--- a/frontend/components/CellInput.js
+++ b/frontend/components/CellInput.js
@@ -421,8 +421,34 @@ export const CellInput = ({
         }, [on_change])
     )
 
-    useLayoutEffect(function cellinput_setup_codemirror() {
+    const [show_static_fake, set_show_static_fake] = useState(true)
+
+    // useEffect(() => {
+    //     setTimeout(() => {
+    //         set_show_static_fake(false)
+    //     }, Math.random() * 6000)
+    // }, [])
+
+    useLayoutEffect(() => {
+        if (!show_static_fake) return
+        let node = dom_node_ref.current
+        if (node == null) return
+        let observer = new IntersectionObserver((e) => {
+            if (e.some((e) => e.isIntersecting)) {
+                set_show_static_fake(false)
+                observer.disconnect()
+            }
+        })
+
+        observer.observe(node)
+        return () => observer.disconnect()
+    }, [])
+
+    useLayoutEffect(() => {
+        if (show_static_fake) return
         if (dom_node_ref.current == null) return
+
+        console.log("Rendering cell input", cell_id)
 
         const keyMapSubmit = (/** @type {EditorView} */ cm) => {
             autocomplete.closeCompletion(cm)
@@ -843,7 +869,7 @@ export const CellInput = ({
                 lines_wrapper_resize_observer.unobserve(lines_wrapper_dom_node)
             }
         }
-    }, [])
+    }, [show_static_fake])
 
     useEffect(() => {
         if (newcm_ref.current == null) return
@@ -922,6 +948,7 @@ export const CellInput = ({
 
     return html`
         <pluto-input ref=${dom_node_ref} class="CodeMirror" translate=${false}>
+            ${show_static_fake ? html`<${StaticCodeMirrorFaker} value=${remote_code} />` : null}
             <${InputContextMenu}
                 on_delete=${on_delete}
                 cell_id=${cell_id}
@@ -1088,3 +1115,31 @@ const InputContextMenuItem = ({ contents, title, onClick, setOpen, tag }) =>
             <span class=${`${tag} ctx_icon`} />${contents}
         </button>
     </li>`
+
+const StaticCodeMirrorFaker = ({ value }) => {
+    const lines = value.split("\n").map((line, i) => html`<div class="awesome-wrapping-plugin-the-line cm-line" style="--indented: 0px;">${line}</div>`)
+
+    return html`
+        <div class="cm-editor ͼ1 ͼ2 ͼ4 ͼ4z">
+            <div tabindex="-1" class="cm-scroller">
+                <div class="cm-gutters" aria-hidden="true">
+                    <div class="cm-gutter cm-lineNumbers"></div>
+                </div>
+                <div
+                    spellcheck="false"
+                    autocorrect="off"
+                    autocapitalize="off"
+                    translate="no"
+                    contenteditable="false"
+                    style="tab-size: 4;"
+                    class="cm-content cm-lineWrapping"
+                    role="textbox"
+                    aria-multiline="true"
+                    aria-autocomplete="list"
+                >
+                    ${lines}
+                </div>
+            </div>
+        </div>
+    `
+}