Render the code as plaintext (with some fake codemirror classes) until the input is scrolled into view, instead of creating the codemirror on page load.

This makes the page load faster (LCP from 720ms to 560ms) and the page becomes interactive faster (busy CPU time from 2070ms to 1020ms). 

The number of elements in the page goes down (from 8409 to 5206, or 4506 when also skipping folded cells).

This also makes SSR easier if we ever want to do that in the future.

# Before
<img width="746" alt="Scherm­afbeelding 2024-04-08 om 12 10 28" src="https://github.com/fonsp/Pluto.jl/assets/6933510/72ae839b-856b-46d1-acf2-da3e29fd8d6d">

# After
<img width="740" alt="Scherm­afbeelding 2024-04-08 om 12 10 44" src="https://github.com/fonsp/Pluto.jl/assets/6933510/5e05a7d9-7375-4aca-9015-5d778c6ed676">


# TODO
- [x] detect print
- [ ] skip optimization for the first 5 cells
- [ ] start loading in background when idle
- [x] Cross-navigating cells with arrow keys should still work
- [x] Investigate skipping render for folded cells
- [x] frontend tests